Update .gitignore, versioning, and toolbar icon logic
- Enhanced `.gitignore` to ignore additional build artifacts and temporary files, while removing outdated IntelliJ entries.
- Updated `CHANGELOG.md` to reflect version change to `3.0.2 (beta.3)` and included various updates and fixes.
- Updated commit reference in the `FauxCore` subproject.
- Changed version in `IconicFramework.csproj` from `3.0.2-beta.2` to `3.0.2-beta.3`.
- Modified event subscription in `ConfigMenu.cs` for better rendering event handling.
- Introduced `ToolbarIconOption.cs` to manage toolbar icon behavior, including rendering and interaction logic.
