+++ .gitignore
@@ -1,36 +1,14 @@
-# Common IntelliJ Platform excludes
-
-# User specific
-**/.idea/**/workspace.xml
-**/.idea/**/tasks.xml
-**/.idea/shelf/*
-**/.idea/dictionaries
-**/.idea/httpRequests/
-
-# Sensitive or high-churn files
-**/.idea/**/dataSources/
-**/.idea/**/dataSources.ids
-**/.idea/**/dataSources.xml
-**/.idea/**/dataSources.local.xml
-**/.idea/**/sqlDataSources.xml
-**/.idea/**/dynamic.xml
-
-# Rider
-# Rider auto-generates .iml files, and contentModel.xml
-**/.idea/**/*.iml
-**/.idea/**/contentModel.xml
-**/.idea/**/modules.xml
-
 *.suo
 *.user
 .vs/
+.idea/
 [Bb]in/
 [Oo]bj/
 _UpgradeReport_Files/
 [Pp]ackages/
+releases/
 
 Thumbs.db
 Desktop.ini
 .DS_Store
 IconicFramework/manifest.json
-releases/
