+++ IconicFramework/IconicFramework.csproj
@@ -6,7 +6,7 @@
 	<PropertyGroup>
 		<Name>Iconic Framework</Name>
 		<Description>Adds shortcut icons to vanilla and mod functions.</Description>
-		<Version>3.0.2-beta.2</Version>
+		<Version>3.0.2-beta.3</Version>
 		<RootNamespace>LeFauxMods.IconicFramework</RootNamespace>
 		<UniqueId>furyx639.ToolbarIcons</UniqueId>
 		<UpdateKeys>Nexus:11026, CurseForge:871834, ModDrop:1547451, GitHub:LeFauxMatt/IconicFramework</UpdateKeys>
